fixed create and delete user scripts so a user can't be added more than once

+++ server/ssh-server/create-user.sh
@@ -9,19 +9,25 @@ mkdir -p /data/$2/.ssh
 touch /data/$2/.ssh/authorized_keys
 
 # Configure SSHD for SCP/SFTP-only Access
-echo "" >> /etc/ssh/sshd_config && \
-echo "Match User $1" >> /etc/ssh/sshd_config && \
-echo "ChrootDirectory /data/$2" >> /etc/ssh/sshd_config
+# Look for an entry for this user in sshd_config
+FOUND=$(grep "Match User $1\$" /etc/ssh/sshd_config | wc -l)
 
-# Find the PID of sshd
-SSHD_PID=$(pidof /usr/sbin/sshd)
+# Only add the user to sshd_config if they are not found
+if [ $FOUND -eq 0 ]; then
+    echo "" >> /etc/ssh/sshd_config && \
+    echo "Match User $1" >> /etc/ssh/sshd_config && \
+    echo "ChrootDirectory /data/$2" >> /etc/ssh/sshd_config
+    # Find the PID of sshd
+    SSHD_PID=$(pidof /usr/sbin/sshd)
 
-# Check if SSHD is running
-if [ -n "$SSHD_PID" ]; then
-    # Send SIGHUP to SSHD to gracefully reload it
-    kill -HUP $SSHD_PID
+    # Check if SSHD is running
+    if [ -n "$SSHD_PID" ]; then
+        # Send SIGHUP to SSHD to gracefully reload it
+        kill -HUP $SSHD_PID
+    else
+        # Start SSHD if it's not running
+        /usr/sbin/sshd -D -e
+    fi
 else
-    # Start SSHD if it's not running
-    /usr/sbin/sshd -D -e
+    echo "user was already in sshd_confi, skipping"
 fi
-
